Test using a simulated server (struct fakeServer)
Add bufio.Writer so that NewClient() can work over an io.Pipe().
The problem was that WriteWord("") was calling Write() twice: once for
the length, and then again for the word itself. On the other end,
readWord() was calling Read() for the length, and since the length is
zero, returning immediately, without calling Read() a second time.
io.Pipe() will wait on Write() until Read() is called, even when Write()
is given a zero-length slice, at least on go1.6.2 windows/amd64. That
resulted in a deadlock in proto_test.go, with 2 goroutines stuck on Read().